Data base

+++ SSoftProject/Models/Event.swift
@@ -6,7 +6,7 @@
 //
 
 struct Event: Codable, Hashable {
-    let id: Int
+    let id: Int64
     let image: String
     let title: String
     let details: String
